Fixes angular scaffolding
Schematics have an issue where depending on the version of schematics in use there are different errors when one attempts to compose a scaffold of multiple template roots - in this case, a template for the manifest and another for the component.

With schematics 0.6 (angular 6) this results in an undefined entries. With schematics 7, it results in an infinite loop.

This commit works around that by not rebasing the manifest template and switching the manifest template to have a full path to the file. Doing this removes the ability to variableize the `manifestPath` but it does allow the scaffold to actually work. This also updates schematics to refer to schematics for angular 7+, which shipped with GA.