Fixed angular graphql SSR in headless mode issue
If the GraphQL request was proxied during SSR through the headless proxy URL (i.e. http://localhost:proxyport), it would have `accept-encoding: gzip` added to it automatically. Problem is, the SSR fetch client does not know how to decompress HTTP responses so that caused SSR to fail as it tried to parse some 'real zipped BSON' :P

This fixes that by only setting the accept encoding header from the proxy IF the incoming request also had one. Yes, this is imperfect - it doesn't account for say 'Accept-Encoding: gzip' - but it's fast and it works for modern browsers.