Prevent reloading Angular route when only hash changes
Refer to #132 - previously the change of the hash would change the route URL but the comparison to refresh the page was not triggered including the current location hash. This meant it was treated as a page refresh situation - but it would never match the current location without the hash so it went into a loop instead.