Updated en,md
Added a Note: When using SXA with JSS ensure the site and app node are commented in the Sitecore `.config` file in `/sitecore/config`. Since this will be handled by SXA Site manager. We faced this issue and took two weeks with Sitecore to figure out the issue. https://sitecore.stackexchange.com/questions/29221/using-page-designs-or-partial-designs-in-sxa-jss-site