[Typescript] Bug fix to allow custom components
Support for custom components in field-factory.tsx I am trying to add a custom react component to the jss-react-forms and relying on typescript.
The method definition only allows of type FieldTypes Enum and I dont have an override to pass my custom component implementation. Interally its a map that takes string and react component for that type.

I have added string so it can be FieldTypes for OOB field and custom field can be added with the {guid} string.