Avoid FEAAS Stylesheets parsing multiple times (#1320)
* WIP: Avoid FEAAS Stylesheets parsing multiple times

FEAAS Components require stylesheet of their libraries to run. Multiple components from the same library can be used on a single page. To ensure that stylesheets start loading as soon as possible, I suggested including stylesheet tag for each component. However with this approach even if stylesheet is loaded once (browser deduplicates requests) it's parsed and initialized multiple times (which can be seen in chrome dev console, having css rules multiplied). 

The solution to this problem is to use `rel="preload"` which will initiate loading during page parsing, so by the time JS is loaded, the static files are already pre-fetched. We also do it for preloading HTML and JS. Combined with SSG, this makes the files load as soon as possible, and still on demand.

* Update FEaaSComponent.tsx

Self-close the stylesheet

* Update clientside

* Update clientside

* Update clientside

* Add tests

* Fix extension

* Remove console logs

---------

Co-authored-by: Adam Brauer <400763+ambrauer@users.noreply.github.com>