Fix: prevent Cannot use ‘in’ operator to search for ‘editable’ in
To reproduce the issue "Cannot use the 'in' operator to search for 'editable' in [SOME-GUIDS]," follow these steps:

* Create a Page Template.
* Create a Page Design.
* Create an Instance of a page.
* Override the Page Design from the Dropdown in the Content Editor.
* From a multilist, select the page which has the Page Design overwritten.

Upon completing these steps, you may encounter the error message. This issue could be due to the structure of Partial Designs, which should ideally follow the {value, editable} structure instead of the provided structure:
{
  "fields": {
    "PartialDesigns": "SOME GUIDS"}"
  }
}

The solution is to validate if field is type of string then were are on this scenario.